cc/interactions: move functions to separate file
